Move USDSceneWriterTest contents to USDSceneTest
In practice there isn't a clear-cut distinction between reading and writing tests, as many tests do both.
